+++ ast/Cargo.toml
@@ -1,15 +1,11 @@
 [package]
-name = "parser"
+name = "ast"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-lexer = { path = "../lexer" }
-context = { path = "../context" }
-ast = { path = "../ast" }
 dbg-pls = { version = "0.3.5", features = ["derive"] }
-
-[dev-dependencies]
-pretty_assertions = "1.3.0"
+enum-assoc = "0.4.0"
+lexer = { path = "../lexer" }
