+++ ast/src/control_flow.rs
@@ -1,4 +1,4 @@
-use crate::ast::Expr;
+use crate::Expr;
 use dbg_pls::DebugPls;
 
 #[derive(Debug, Clone, PartialEq, DebugPls)]
