+++ ast/src/substitution.rs
@@ -1,4 +1,4 @@
-use crate::ast::group::Subshell;
+use crate::group::Subshell;
 use dbg_pls::DebugPls;
 
 /// A special type of grouping expression that should be substituted
