+++ ast/src/test.rs
@@ -1,11 +1,11 @@
-use crate::ast::Expr;
+use crate::Expr;
 use dbg_pls::DebugPls;
 
 /// a test (`[ ... ]`) expression
 #[derive(Debug, Clone, PartialEq, DebugPls)]
 pub struct Test<'a> {
     ///expression present between brackets
-    pub(crate) expression: Box<Expr<'a>>,
+    pub expression: Box<Expr<'a>>,
 }
 
 ///a not (`! ..`) expression
@@ -13,5 +13,5 @@ pub struct Test<'a> {
 
 pub struct Not<'a> {
     ///the expression after `!`
-    pub(crate) underlying: Box<Expr<'a>>,
+    pub underlying: Box<Expr<'a>>,
 }
