+++ ast/src/variable.rs
@@ -1,4 +1,4 @@
-use crate::ast::Expr;
+use crate::Expr;
 use dbg_pls::DebugPls;
 
 /// A typed variable.
