+++ parser/src/aspects/binary_operation.rs
@@ -1,5 +1,5 @@
-use crate::ast::operation::{BinaryOperation, BinaryOperator};
-use crate::ast::Expr;
+use ast::operation::{BinaryOperation, BinaryOperator};
+use ast::Expr;
 use crate::moves::{bin_op, eox, spaces, word_seps, MoveOperations};
 use crate::parser::{ParseResult, Parser};
 
@@ -155,12 +155,12 @@ mod tests {
     use context::source::Source;
 
     use crate::aspects::binary_operation::BinaryOperationsAspect;
-    use crate::ast::callable::Call;
-    use crate::ast::group::{Parenthesis, Subshell};
-    use crate::ast::operation::BinaryOperation;
-    use crate::ast::operation::BinaryOperator::*;
-    use crate::ast::value::Literal;
-    use crate::ast::Expr;
+    use ast::callable::Call;
+    use ast::group::{Parenthesis, Subshell};
+    use ast::operation::BinaryOperation;
+    use ast::operation::BinaryOperator::*;
+    use ast::value::Literal;
+    use ast::Expr;
     use crate::err::{ParseError, ParseErrorKind};
     use crate::parser::Parser;
