+++ parser/src/aspects/group.rs
@@ -1,7 +1,7 @@
 use lexer::token::{Token, TokenType};
 
-use crate::ast::group::{Block, Parenthesis, Subshell};
-use crate::ast::Expr;
+use ast::group::{Block, Parenthesis, Subshell};
+use ast::Expr;
 use crate::err::ParseErrorKind;
 use crate::moves::{eox, of_type, repeat, repeat_n, spaces, MoveOperations};
 use crate::parser::{ParseResult, Parser};
@@ -131,12 +131,12 @@ impl<'a> Parser<'a> {
 #[cfg(test)]
 mod tests {
     use crate::aspects::group::GroupAspect;
-    use crate::ast::callable::Call;
-    use crate::ast::group::{Block, Subshell};
-    use crate::ast::value::Literal;
-    use crate::ast::value::LiteralValue::{Float, Int};
-    use crate::ast::variable::{TypedVariable, VarDeclaration, VarKind};
-    use crate::ast::Expr;
+    use ast::callable::Call;
+    use ast::group::{Block, Subshell};
+    use ast::value::Literal;
+    use ast::value::LiteralValue::{Float, Int};
+    use ast::variable::{TypedVariable, VarDeclaration, VarKind};
+    use ast::Expr;
     use crate::parser::Parser;
     use context::source::Source;
     use pretty_assertions::assert_eq;
