+++ parser/src/aspects/match.rs
@@ -4,9 +4,9 @@ use lexer::token::TokenType::{
 };
 
 use crate::aspects::literal::LiteralAspect;
-use crate::ast::r#match::MatchPattern::{Literal, Template, VarRef, Wildcard};
-use crate::ast::r#match::{Match, MatchArm, MatchPattern};
-use crate::ast::Expr;
+use ast::r#match::MatchPattern::{Literal, Template, VarRef, Wildcard};
+use ast::r#match::{Match, MatchArm, MatchPattern};
+use ast::Expr;
 use crate::err::ParseErrorKind;
 use crate::moves::{
     aerated, any, blanks, eod, eox, not, of_type, of_types, repeat, MoveOperations,
@@ -204,14 +204,14 @@ mod tests {
     use context::source::Source;
     use pretty_assertions::assert_eq;
 
-    use crate::ast::callable::Call;
-    use crate::ast::group::Subshell;
-    use crate::ast::operation::{BinaryOperation, BinaryOperator};
-    use crate::ast::r#match::{Match, MatchArm, MatchPattern};
-    use crate::ast::test::Test;
-    use crate::ast::value::{Literal, TemplateString};
-    use crate::ast::variable::{TypedVariable, VarDeclaration, VarKind, VarReference};
-    use crate::ast::Expr;
+    use ast::callable::Call;
+    use ast::group::Subshell;
+    use ast::operation::{BinaryOperation, BinaryOperator};
+    use ast::r#match::{Match, MatchArm, MatchPattern};
+    use ast::test::Test;
+    use ast::value::{Literal, TemplateString};
+    use ast::variable::{TypedVariable, VarDeclaration, VarKind, VarReference};
+    use ast::Expr;
     use crate::err::ParseError;
     use crate::err::ParseErrorKind::Unexpected;
     use crate::parse;
