+++ parser/src/aspects/redirection.rs
@@ -1,6 +1,6 @@
 use crate::aspects::call::CallAspect;
-use crate::ast::callable::{Pipeline, Redir, RedirFd, RedirOp, Redirected};
-use crate::ast::Expr;
+use ast::callable::{Pipeline, Redir, RedirFd, RedirOp, Redirected};
+use ast::Expr;
 use crate::err::ParseErrorKind;
 use crate::moves::{eox, like, next, of_type, of_types, spaces, MoveOperations};
 use crate::parser::{ParseResult, Parser};
@@ -177,10 +177,10 @@ mod test {
     use pretty_assertions::assert_eq;
 
     use crate::aspects::call::CallAspect;
-    use crate::ast::callable::{Call, Redir, RedirFd, RedirOp, Redirected};
-    use crate::ast::group::Block;
-    use crate::ast::value::Literal;
-    use crate::ast::Expr;
+    use ast::callable::{Call, Redir, RedirFd, RedirOp, Redirected};
+    use ast::group::Block;
+    use ast::value::Literal;
+    use ast::Expr;
     use crate::parse;
     use crate::parser::Parser;
