+++ parser/src/aspects/var_reference.rs
@@ -1,8 +1,8 @@
 use context::source::try_join_str;
 use lexer::token::TokenType::*;
 
-use crate::ast::variable::VarReference;
-use crate::ast::Expr;
+use ast::variable::VarReference;
+use ast::Expr;
 use crate::err::ParseErrorKind;
 use crate::moves::{like, of_type, repeat, MoveOperations};
 use crate::parser::{ParseResult, Parser};
@@ -52,9 +52,9 @@ impl<'a> VarReferenceAspect<'a> for Parser<'a> {
 #[cfg(test)]
 mod tests {
     use crate::aspects::substitution::SubstitutionAspect;
-    use crate::ast::value::TemplateString;
-    use crate::ast::variable::VarReference;
-    use crate::ast::Expr;
+    use ast::value::TemplateString;
+    use ast::variable::VarReference;
+    use ast::Expr;
     use crate::err::{ParseError, ParseErrorKind};
     use crate::parse;
     use crate::parser::{ParseResult, Parser};
