+++ parser/src/lib.rs
@@ -8,7 +8,6 @@ use crate::parser::Parser;
 
 ///! The parser crate contains the parser for the Moshell scripting language.
 mod aspects;
-pub mod ast;
 mod cursor;
 pub mod err;
 mod moves;
