+++ parser/tests/err.rs
@@ -1,6 +1,6 @@
 use context::source::Source;
-use parser::ast::callable::Call;
-use parser::ast::Expr;
+use ast::callable::Call;
+use ast::Expr;
 use parser::err::{ParseError, ParseErrorKind, ParseReport};
 use parser::parse;
